Update README.md
This repository is to share my workings as a Web Development Intern in the Virtual Internship Program at Oasis Infobytes which is held in this Month of October 2023. I would like to thank the team of Oasis Infobytes for giving me this opportunity to enhance my skills like HTML,CSS, Javascript and work on such an amazing projects.